+++ src/directives/index.ts
@@ -1,3 +1,4 @@
 export * from './normalStateTemplate/normalStateTemplate.directive';
 export * from './optionTemplate/optionTemplate.directive';
 export * from './selectAbsolute/selectAbsolute.directive';
+export * from './withDirectAccess/withDirectAccess.directive';