+++ src/directives/withDirectAccess/withDirectAccess.directive.ts
@@ -1,0 +1,14 @@
+import {Directive} from '@angular/core';
+
+//TODO: finish
+
+/**
+ * Directive that allows direct access to select properties using inputs, outputs
+ */
+@Directive(
+{
+    selector: '[withDirectAccess]',
+})
+export class WithDirectAccess
+{
+}