+++ src/plugins/normalState/simpleNormalState/simpleNormalState.component.ts
@@ -41,11 +41,4 @@ export class SimpleNormalStateComponent implements NormalState
     public initOptions(): void
     {
     }
-
-    /**
-     * @inheritdoc
-     */
-    public invalidateVisuals(): void
-    {
-    }
 }