+++ src/plugins/popup/simplePopup/simplePopup.component.ts
@@ -41,11 +41,4 @@ export class SimplePopupComponent implements Popup
     public initOptions(): void
     {
     }
-
-    /**
-     * @inheritdoc
-     */
-    public invalidateVisuals(): void
-    {
-    }
 }